fix: changed baseFee from Int to Float (#43)
fixes https://github.com/apotdevin/thunderhub/issues/42

+++ src/graphql/mutations/updateFees.ts
@@ -5,7 +5,7 @@ export const UPDATE_FEES = gql`
     $auth: authType!
     $transactionId: String
     $transactionVout: Int
-    $baseFee: Int
+    $baseFee: Float
     $feeRate: Int
   ) {
     updateFees(
